Initial impl of StepRunner API server
Most of this code comes from cmd/ci/ci.go. There/s probably some
consolidation to be made here.

This is a working but thin implementation of the APIs. Specifically jobs
are not executed in a streaming fashion. The slice of jobs are crammed
into a `StepDefinition`, executed in one shot, and the entire result is
returned at the end. I took this shortcut so I could just reuse
`execution.Run()`; we'll need an `execution.RunStep()` to be able to run
and stream back results one step at at a time.

It probably requires more/better error handling too.