Capture a step's stdout
We want to capture the step execution's stdout to stream back to the job log.
For now, KIS and add a field to the FollowResponse message. This will
result in a sort of chunked streaming, where a steps output is returned
in one shot instead of "live streamed".

Also combing stdout and stderr for now to KIS.