Add WorkDir to RunRequest and propagate it down to command execution
This WorkDir is propagated down to `runSteps()`, `runStep()` ultimately
`runExec()`, and `takes precedence over the step-definition WorkDir.
There is a single work-for the entire `RunRequest` execution context; if
we need a separate work-dir for each step this worn't work.

This lets the client specify in what dir the commands ultimately
executed will be run.