Replace Cobra with go-arg
go-arg is much simpler and lighter weight that cobra. The benefits of
go-arg won't be apparent in this commit because the `ci` subcommand has
no arguments, but dealing with options, including positional ones, is
slick. It's all just struct tags!

Subcommands are just struct-pointers in the top-level struct. All the
magic happens in `arg.MustParse()`.

The top-level `run` function is one way to handle subcommands.
`Subcommand()` returns the struct corresponding to the subcommand that
was invoked. Another, more explicit and verbose way is like so:

func run() error {
    app := App{}
    p := arg.MustParse(&app)

    switch {
    case app.CI != nil:
            return app.CI.Run()
    default:
            p.WriteHelp(os.Stderr)
    }
    return nil
}

Cobra
    Step Runner executes a series of CI steps

    Usage:
    step-runner [command]

    Available Commands:
    ci          Run steps in a CI environment variable STEPS
    completion  Generate the autocompletion script for the specified shell
    help        Help about any command

    Flags:
    -h, --help   help for step-runner

    Use "step-runner [command] --help" for more information about a command.

go-arg
    Step Runner executes a series of CI steps
    Usage: step-runner <command> [<args>]

    Options:
    --help, -h             display this help and exit

    Commands:
    ci                     Run steps in a CI environment variable STEPS