Add jobs package and type
Package jobs (and specifically the type Job) implements managing the
lifecycle of a steps run request. It does thing like setting temp paths,
tracking the state of job execution, capturing results and errors,
handling job cancellation, and cleaning up resources when the job is
finished. 

It is intended to be used by the step-runner gRPC API. This package is a
work in process since only one of the 4 APIs have been implemented; as
the API grows so will the responsibilities of this package.