Implement Run API
This is an initial implementation of the Run API. Most of the fun is in
the Run function, which parses the step in the request, prepares the
execution environment (job and environment variables, work and temp
dirs), and finally initiates concurrent execution of the request.

The run method actually runs the request, captures the output, and
updates the associates Job.