Formalize bad job-id error
A number of the APIs will have to look up the job-id in the request and
handle when that id does not exist. Formalizing it ensures a consistent
message. Unfortunately the type is not retained across the gRPC call, to
the API caller can't do things like `errors.Is( err, &errNoJobID)`.