In-memory Streamer
As implementation of the Streamer idea, that buffers data in a
slice. We'll use this later to enable streaming of StepResults.

The notable part here is the use of `sync.Cond` to synchronise the
writer and reader(s). On the follower side, `Cond.Wait()` will stop the
calling gorouteine(s) and thus pause when all StepResults collected so
far have been forwarded. On the Writer side calling `Cond.Broadcast()`
after a StepResult has been collected will wake up all stopped
goroutines.

The Follower(s) must be told when to stop via the Stop() method.