ExecutionRunHook type
This interface will allow users of runner.Execution to invoke arbitrary
code when Run() is called. We'll use this later to capture StepResults
as they are created for each sub-step to enable proper StepResult
streaming via the API.

Currently this one has two methods, but more can be added if
needed/desired. Also, the argument to PostEnter() might not be the best.

I've also added a no-op implementation to keep the call-site simple.