File Streamer
This type is very similar but not identical to the in-memory streamer
introduced in #65. The backing store is a file instead of memory though,
and the data type is limited to `[]byte` instead of being generic.

Note that the Write method implements io.Writerl; this will be necessary
later on to transparently add things like log formatting and secret
masking.