Extended client
This client build on the basic client and adds handling of the complete
life-cycle of a run request, including (eventually) reconnection and
follow on recoverable disconnect errors.

This is a well-behaved behaved client for the step-runner gRPC service.
The primary entry point is RunAndFollow, which manages the entire
lifecycle of the steps job, including calling the Close() API to
potentially cancel a running job, and release all resources associated
with it.

Note that if it is cancelled or times out, the context passed to Follow
will ultimately cancel the job on the server side too since it will
cause the Close() API to be called.

Reconnection handling is not implemented here, but there is room made
for it by accepting a DialFunc in New. Adding reconnection will come
later.