Proxy implementation
Proxy connects the specified read and writer with the specified
connection. It copies everything from the source to the connection's
`io.Writer`, and the connection's `io.Reader` to the sink.

Under typical usage the connection will be to the steps gRPC service
listening on api.DefaultSocketPath, and the source and sink will be the
stdin and stdout (respectively) of a streaming-text based protocol like
`ssh` or `docker exec`.