Basic client
Split the existing client into basic and extended clients. This is the
basic 1/2. This client is low-level and more or less 1:1 mapping to the
raw gRPC API. It abstracts all proto types except for StepResult. Use
this API if you want full control over every stage of running, following
and closing a job.