Simplify cleaning up job on Run error or cancellation
If the Run request fails for any reason, we want to ensure `job.Close()`
is called to clean up and free resources associated with the job.
Previously `job.Close()` had to be called explicitly at every non-nil
error return. This change makes it automatic.

This is a bit tricky to tests since it's not possible to make any of the
functions called in Run to return a non nil error. About all we can do
it ensure the passed in context is canceled to trigger the final if
guard.