Make a step's SpecDef.Dir the BuildDir
Otherwise loading local steps fails.

This required changing a number of tests. Without setting SpecDef.Dir,
the tests by chance were in the right working dir to load the local
steps. Now that we set SpecDef.Dir, we have to set the RunRequest's
WorkDir (or BuildDir) to the test's working dir for local step loading
to work.

Some steps pollute the WorkDir so we have to clean that too.