Close only the Read or Write half of the network connection
If we close the entire connection when exiting either of the read/write
loop, we get back the following error from the corresponding io.Copy():

    Error: proxying "/tmp/step-runner.sock" to stdout: *fs.PathError : write /dev/stdout: use of closed network connection

This error (rightly) appears in the runner job log trace, and implies an
error occurred when in really this is the normal exit procedure for the
proxy. Closing only the Write/Read half results in a EOF, which
io.Copy() handles by returning a nil error.

Note also that we are capturing the Close errors in the defered function
and returning them to the caller.