Don't print out version when executing the proxy command
Because the proxy reads from stdin and write to stdout on one side of
the connection, printing anything to stdout breaks the whole thing. The
relevant error message is.

    executing step request: running job request: rpc error: code = Unavailable desc = connection error: desc = "error reading server preface: http2: frame too large"

To fix the above I had to move printing of the version to the root
command, where I can detect what subcommand is being run. To to the
latter I had to make some changes to how ldflags are handled.