Refactor/Fix Job.computeFinalStatus()
When a job is cancelled (by calling `Job.Close()`) or times out (both of
which cancel the context passed to `exec.CommandContext()`), the returned
error can:
 * be one of context.Cancelled or context.DeadlineExceeded.
 * be another error type that ends with the string "signal: killed".

In both cases the `StepResult_Status` returned by `Step.Run()` is
`failure`, but we want it to be `cancelled`. Since the latter can also
happen when the process is otherwise killed (e.g. OOM killer), so we
have to also check that the context was actually cancelled.