Roll Job.Finish() into Job.Run()
Move everything that was in `Finish()` into `Run()`. This means that
what was in finish can now only be called from one place (`Run()`).
Instead of Calling `Finish()` directly, `Close()` now waits for `Run()`
to exit using a channel. Because `Close()` block on a channel, we wrap
it in a `sync.Once` to prevent a deadlock when calling `Close()`
multiple times.

Also update the affected unit tests.