Add bootstrap command
This command copies the executing step-runner binary to the specified
destination dir. The destination file will be `step-runner`.

We'll use this later/elsewhere to inject the step-runner binary in to
the target environment by making the container entrypoint the bootstrap
command, with the destination the volume mounted by both the build
container and the step-runner container.