Remove unused Go module dependencies
Over in the runner project, updating to the latest steps version was
pulling in an update of various docker libraries. The jump was several;
MAJOUR versions, which was breaking all manner of things. AFAICT it
turns out the docker dependencies here were triggering that, bu those
modules are not actually used here.