Serve command accepts a unix socket path
I need this primarily to spawn step-runner services locally from unit
tests elsewhere. I've made this a positional, optional, unnamed arg
since we don't need/want to advertise it's existence widely yet.

Note that I have deliberately not added this same option to the `proxy`
command. I don't need it there now, there's no harm in omitting it, and
it'll be easy to add if we need it.